Rename parse date utility function package to utils
