+++ internal/utils/parse-date.go
@@ -1,4 +1,4 @@
-package parsers
+package utils
 
 import "time"
